requirements.txt
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.40
botocore==1.31.40
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
waitress==2.1.2
Werkzeug==2.3.7
